Moved `EuiHeaderNotification`
to generic `EuiNotificationBadge`
