+++ src/components/badge/_index.scss
@@ -1,2 +1,3 @@
 @import 'badge';
 @import 'beta_badge/index';
+@import 'notification_badge/index';
