+++ src/components/badge/index.js
@@ -5,3 +5,7 @@ export {
 export {
   EuiBetaBadge,
 } from './beta_badge';
+
+export {
+  EuiNotificationBadge,
+} from './notification_badge';
