+++ src/components/badge/notification_badge/_badge_notification.scss
@@ -1,0 +1,14 @@
+.euiNotificationBadge {
+  display: inline-block;
+  border-radius: $euiBorderRadius;
+  background: $euiColorAccent;
+  color: lightOrDarkTheme($euiColorEmptyShade, $euiColorFullShade);
+  font-size: $euiFontSizeXS;
+  line-height: $euiSize;
+  height: $euiSize;
+  min-width: $euiSize;
+  padding-left: $euiSizeXS;
+  padding-right: $euiSizeXS;
+  vertical-align: middle;
+  text-align: center;
+}
