+++ src/components/badge/notification_badge/_index.scss
@@ -1,0 +1,1 @@
+@import 'badge_notification';
