+++ src/components/badge/notification_badge/badge_notification.test.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { render } from 'enzyme';
+import { requiredProps } from '../../../test';
+
+import { EuiNotificationBadge } from './badge_notification';
+
+describe('EuiNotificationBadge', () => {
+  test('is rendered', () => {
+    const component = render(
+      <EuiNotificationBadge {...requiredProps} />
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+});
