+++ src/components/badge/notification_badge/index.js
@@ -1,0 +1,3 @@
+export {
+  EuiNotificationBadge,
+} from './badge_notification';
