+++ src/components/header/_header.scss
@@ -7,15 +7,3 @@
   background: $euiHeaderBackgroundColor;
   border-bottom: $euiBorderThin;
 }
-
-  .euiHeader__notification {
-    display: inline-block;
-    border-radius: $euiBorderRadius;
-    background: $euiHeaderNotificationBackgroundColor;
-    color: $euiColorEmptyShade;
-    font-size: $euiFontSizeXS;
-    line-height: $euiSize;
-    height: $euiSize;
-    min-width: $euiSize;
-    vertical-align: middle;
-  }
