--- src/components/header/_header_notification.scss
@@ -1,12 +1,0 @@
-
-.euiHeaderNotification {
-  display: inline-block;
-  border-radius: $euiBorderRadius;
-  background: $euiHeaderNotificationBackgroundColor;
-  color: $euiColorEmptyShade;
-  font-size: $euiFontSizeXS;
-  line-height: $euiSize;
-  height: $euiSize;
-  min-width: $euiSize;
-  vertical-align: middle;
-}
