+++ src/components/index.js
@@ -14,6 +14,7 @@ export {
 export {
   EuiBadge,
   EuiBetaBadge,
+  EuiNotificationBadge,
 } from './badge';
 
 export {
@@ -141,7 +142,6 @@ export {
   EuiHeaderLink,
   EuiHeaderLinks,
   EuiHeaderLogo,
-  EuiHeaderNotification,
   EuiHeaderSection,
   EuiHeaderSectionItem,
   EuiHeaderSectionItemButton,
